# src/main/resources/static/openapi.yaml
openapi: "3.0.2"
info:
  title: Freelancer Web Services API
  version: "0.1"
  description: >
    The Freelancer Web Services API is a RESTful API that allows the registration and retrieval of freelancers.
    It is a simple API that serves as a reference for the development of RESTful APIs.
servers:
  - url: http://localhost:8080
    description: Local server
paths:
...
  /freelancers:
    get:
      tags:
        - Freelancers
      description: Returns a paginated list of freelancers.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: 0-based page number to list
        - in: query
          name: size
          schema:
            type: integer
          required: false
          description: Number of elements per page
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                type: object
                required:
                  - _links
                properties:
                  _embedded:
                    type: object
                    required:
                      - freelancers
                    properties:
                      freelancers:
                        type: array
                        items:
                          $ref: "#/components/schemas/Freelancer"
                  _links:
                    type: object
                    required:
                      - self
                      - register
                    properties:
                      self:
                        $ref: "#/components/schemas/Link"
                      register:
                        $ref: "#/components/schemas/Link"
                      first:
                        $ref: "#/components/schemas/Link"
                      prev:
                        $ref: "#/components/schemas/Link"
                      next:
                        $ref: "#/components/schemas/Link"
                      last:
                        $ref: "#/components/schemas/Link"
                  page:
                    $ref: "#/components/schemas/Page"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags:
        - Freelancers
      description: Registers a new freelancer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Freelancer"
      responses:
        "201":
          description: Freelancer was registered
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
  /freelancers/{id}:
    get:
      tags:
        - Freelancers
      description: Retrieves a freelancer by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a freelancer already registered.
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
  /freelancers/{id}/address:
    patch:
      tags:
        - Freelancers
      description: Updates the address of a registered freelancer.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a freelancer already registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Address"
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
  /freelancers/{id}/nicknames:
    patch:
      tags:
        - Freelancers
      description: Updates the nicknames of a registered freelancer.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a freelancer already registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
  /freelancers/{id}/wage:
    patch:
      tags:
        - Freelancers
      description: Updates the wage of a registered freelancer.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a freelancer already registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Wage"
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
  /freelancers/{id}/communication-channels:
    post:
      tags:
        - Freelancers
      description: Adds communication channels to a registered freelancer.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a freelancer already registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunicationChannel"
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
  /freelancers/{id}/communication-channels/{channelId}:
    delete:
      tags:
        - Freelancers
      description: Removes communication channels from a registered freelancer.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Identifier of a freelancer already registered.
        - in: path
          name: channelId
          schema:
            type: string
          required: true
          description: Identifier of a communication channel already registered.
      responses:
        "200":
          description: Success
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        default:
          $ref: "#/components/responses/Error"
...
components:
  schemas:
    Freelancer:
      type: object
      description: Information about a freelancer
      properties:
        id:
          type: string
          nullable: true
          description: Unique identifier for the freelancer
          example: "fa8508ed-8b7b-4be7-b372-ac1094c709b5"
        firstName:
          type: string
          description: First name of the freelancer
          example: "Tony"
        middleName:
          type: string
          nullable: true
          description: Optional middle name of the freelancer
          example: "E"
        lastName:
          type: string
          description: Last name of the freelancer
          example: "Stark"
        gender:
          type: string
          description: Gender of the freelancer
          example: "MALE"
        birthDate:
          type: string
          format: date
          description: Birth date of the freelancer
          example: "1970-05-29"
        address:
          $ref: "#/components/schemas/Address"
        wage:
          $ref: "#/components/schemas/Wage"
        nicknames:
          type: array
          description: List of nicknames for the freelancer
          items:
            type: string
          example: ["Iron Man"]
        communicationChannels:
          type: array
          description: List of communication channels for the freelancer
          items:
            $ref: "#/components/schemas/CommunicationChannel"
        _links:
          type: object
          description: Links for subresources and actions available for the freelancer
          required:
            - self
            - changeAddress
          properties:
            collection:
              $ref: "#/components/schemas/Link"
            self:
              $ref: "#/components/schemas/Link"
            changeAddress:
              $ref: "#/components/schemas/Link"
            updateNicknames:
              $ref: "#/components/schemas/Link"
            updateWage:
              $ref: "#/components/schemas/Link"
            addCommunicationChannel:
              $ref: "#/components/schemas/Link"
            removeCommunicationChannel:
              oneOf:
                - $ref: "#/components/schemas/Link"
                - type: array
                  items:
                    $ref: "#/components/schemas/Link"
      required:
        - firstName
        - lastName
        - birthDate
        - address
        - wage
        - nicknames
        - communicationChannels
    Address:
      type: object
      description: Address information
      properties:
        street:
          type: string
          example: "10880 Malibu Point"
        apartment:
          type: string
          nullable: true
          example: "Apt 1A"
        city:
          type: string
          example: "Malibu"
        state:
          type: string
          example: "CA"
        zipCode:
          type: string
          example: "90265"
        country:
          type: string
          example: "USA"
      required:
        - street
        - city
        - state
        - zipCode
        - country
    Wage:
      type: object
      description: Wage information
      properties:
        amount:
          type: number
          format: float
          description: Wage amount
          example: 5000000.00
        currency:
          type: string
          description: Wage currency
          example: "USD"
      required:
        - amount
        - currency
    CommunicationChannel:
      type: object
      properties:
        id:
          type: string
          nullable: true
          description: Unique identifier for the communication channel
          example: "c7e8f8e7-7f7e-4e7f-8e7f-7e7f8e7f7e7f"
        value:
          type: string
          description: Communication channel value (e.g., email address)
          example: "ironman@avengers.org"
        type:
          type: string
          enum:
            - MOBILE
            - PHONE
            - EMAIL
            - URL
          description: Type of communication channel (e.g., EMAIL)
          example: "EMAIL"
      required:
        - value
        - type
...
    Page:
      type: object
      properties:
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 1
        number:
          type: integer
          example: 0
    Link:
      type: object
      description: Link to a resource or action
      properties:
        href:
          type: string
          description: URI of the resource or action
          example: "/collection-name/resource-id"
        method:
          type: string
          description: HTTP method for the link
          example: "GET"
        templated:
          type: boolean
          description: Whether the URI contains template parameters
          example: false
      required:
        - href
        - method
  responses:
    Error:
      description: >
        Error response formatted as per the Internet Standard
        <a href="https://datatracker.ietf.org/doc/html/rfc7807">Problem Details for HTTP APIs (RFC 7807)</a>.
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type:
                description: A URI reference that identifies the problem type.
                type: string
                format: uri-reference
                example: /probs/validation-error
              title:
                description: A short, human-readable summary of the problem type.
                type: string
                example: Bad Request
              status:
                description: The HTTP status code generated by the origin server for this occurrence of the problem.
                type: integer
                example: 400
              detail:
                description: A human-readable explanation specific to this occurrence of the problem.
                type: string
                example: Invalid request content.
              instance:
                description: >
                  A URI reference that identifies the specific occurrence of the problem.
                  It may or may not yield further information if dereferenced.
                type: string
                format: uri-reference
                example: /freelancers
              errors:
                description: Optional array that describes validation errors.
                type: array
                items:
                  type: object
                  required:
                    - detail
                  properties:
                    attribute:
                      description: >
                        Name of the request body's attribute that caused the problem,
                        if not present the whole body should be considered the cause of the problem.
                      type: string
                      example: firstName
                    detail:
                      description: A human-readable description of the error
                      type: string
                      example: must not be blank
            required:
              - type
              - title
              - status
              - instance